+++ packages/rust/src/specification/contract.rs
@@ -1,6 +1,14 @@
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Default)]
+pub struct ContractPair {
+    #[serde(default)]
+    pub input: Contract,
+    #[serde(default)]
+    pub output: Contract,
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Default)]
 pub struct Contract {
     pub schema: String,
     #[serde(skip_serializing_if = "Option::is_none")]